# business/tecma-digital-platform/followup-3.0/be-followup-v3/openapi/openapi.v1.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Followup 3.0 REST API",
    "version": "1.0.0",
    "description": "API REST per Followup 3.0 (CRM multi-progetto). Gestione clienti, appartamenti, trattative (rent/sell), calendario, workspaces, progetti, audit e report. Autenticazione JWT (login, refresh). In locale il backend è su porta 8080."
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Backend locale (dev)"
    },
    {
      "url": "/v1",
      "description": "Stesso host (proxy)"
    }
  ],
...
  "paths": {
...
    "/ai/approvals": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Approve or reject AI suggestion",
        "responses": {
          "200": {
            "description": "Decision recorded"
          },
          "404": {
            "description": "Suggestion not found"
          }
        }
      }
    },
...
    "/ai/actions/drafts/{id}/decision": {
      "post": {
        "tags": [
          "ai"
        ],
        "summary": "Approve or reject AI action draft",
        "responses": {
          "200": {
            "description": "Draft decision recorded and optionally executed"
          },
          "404": {
            "description": "Draft not found"
          }
        }
      }
    },
...
    "/audit/query": {
      "post": {
        "tags": [
          "audit"
        ],
        "summary": "Query log audit (workspaceId, filtri, paginazione)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista record audit"
          }
        }
      }
    },
    "/audit/entity/{entityType}/{entityId}": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Eventi audit per entità (cliente, appartamento, request)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "entityType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "client",
                "apartment",
                "request"
              ]
            }
          },
          {
            "name": "entityId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista eventi"
          }
        }
      }
    },
...
    "/reports/{reportType}": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Genera report (pipeline, clients-by-status, apartments-by-availability, kpi-summary)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reportType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "pipeline",
                "clients_by_status",
                "apartments_by_availability",
                "kpi_summary"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspaceId": {
                    "type": "string"
                  },
                  "projectIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "dateFrom": {
                    "type": "string"
                  },
                  "dateTo": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dati report"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login or refresh"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Platform consumer API key (scoped by workspace/project)."
      },
      "signatureWebhookAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-signature-webhook-secret",
        "description": "Segreto condiviso per webhook firme (stesso valore di SIGNATURE_WEBHOOK_SECRET). Equivalente: Authorization Bearer <secret>."
      }
    },
...
  }
}